LPS-52265 move *ConverterKeys to module
