+++ modules/portal/portal-ldap/src/com/liferay/portal/ldap/DefaultPortalLDAP.java
@@ -31,10 +31,8 @@ import com.liferay.portal.kernel.util.StringUtil;
 import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portal.model.CompanyConstants;
 import com.liferay.portal.security.exportimport.UserImportTransactionThreadLocal;
-import com.liferay.portal.security.ldap.GroupConverterKeys;
 import com.liferay.portal.security.ldap.LDAPSettingsUtil;
 import com.liferay.portal.security.ldap.PortalLDAP;
-import com.liferay.portal.security.ldap.UserConverterKeys;
 import com.liferay.portal.util.PropsValues;
 
 import java.util.ArrayList;
