+++ modules/portal/portal-ldap/src/com/liferay/portal/ldap/UserConverterKeys.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.security.ldap;
+package com.liferay.portal.ldap;
 
 /**
  * @author Edward Han
