+++ modules/portal/portal-ldap/src/com/liferay/portal/ldap/exportimport/DefaultLDAPToPortalConverter.java
@@ -28,6 +28,9 @@ import com.liferay.portal.kernel.util.PropsKeys;
 import com.liferay.portal.kernel.util.StringPool;
 import com.liferay.portal.kernel.util.StringUtil;
 import com.liferay.portal.kernel.util.Validator;
+import com.liferay.portal.ldap.ContactConverterKeys;
+import com.liferay.portal.ldap.GroupConverterKeys;
+import com.liferay.portal.ldap.UserConverterKeys;
 import com.liferay.portal.model.Contact;
 import com.liferay.portal.model.ContactConstants;
 import com.liferay.portal.model.ListType;
@@ -35,12 +38,9 @@ import com.liferay.portal.model.ListTypeConstants;
 import com.liferay.portal.model.User;
 import com.liferay.portal.security.auth.FullNameGenerator;
 import com.liferay.portal.security.auth.FullNameGeneratorFactory;
-import com.liferay.portal.security.ldap.ContactConverterKeys;
-import com.liferay.portal.security.ldap.GroupConverterKeys;
 import com.liferay.portal.security.ldap.LDAPGroup;
 import com.liferay.portal.security.ldap.LDAPToPortalConverter;
 import com.liferay.portal.security.ldap.LDAPUser;
-import com.liferay.portal.security.ldap.UserConverterKeys;
 import com.liferay.portal.service.ListTypeServiceUtil;
 import com.liferay.portal.service.ServiceContext;
 import com.liferay.portal.service.persistence.ContactUtil;
