+++ modules/portal/portal-ldap/src/com/liferay/portal/ldap/exportimport/DefaultPortalToLDAPConverter.java
@@ -25,17 +25,17 @@ import com.liferay.portal.kernel.util.PropsKeys;
 import com.liferay.portal.kernel.util.StringBundler;
 import com.liferay.portal.kernel.util.StringPool;
 import com.liferay.portal.kernel.util.Validator;
+import com.liferay.portal.ldap.GroupConverterKeys;
+import com.liferay.portal.ldap.UserConverterKeys;
 import com.liferay.portal.model.Contact;
 import com.liferay.portal.model.Image;
 import com.liferay.portal.model.User;
 import com.liferay.portal.model.UserGroup;
 import com.liferay.portal.security.exportimport.UserOperation;
-import com.liferay.portal.security.ldap.GroupConverterKeys;
 import com.liferay.portal.security.ldap.LDAPSettingsUtil;
 import com.liferay.portal.security.ldap.Modifications;
 import com.liferay.portal.security.ldap.PortalLDAPUtil;
 import com.liferay.portal.security.ldap.PortalToLDAPConverter;
-import com.liferay.portal.security.ldap.UserConverterKeys;
 import com.liferay.portal.security.pwd.PasswordEncryptorUtil;
 import com.liferay.portal.service.ImageLocalServiceUtil;
 import com.liferay.portlet.expando.model.ExpandoBridge;
