+++ modules/portal/portal-ldap/src/com/liferay/portal/ldap/exportimport/LDAPUserExporterImpl.java
@@ -19,13 +19,13 @@ import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.util.PrefsPropsUtil;
 import com.liferay.portal.kernel.util.PropsKeys;
 import com.liferay.portal.kernel.workflow.WorkflowConstants;
+import com.liferay.portal.ldap.GroupConverterKeys;
 import com.liferay.portal.ldap.PortalLDAPContext;
 import com.liferay.portal.model.Contact;
 import com.liferay.portal.model.User;
 import com.liferay.portal.model.UserGroup;
 import com.liferay.portal.security.exportimport.UserExporter;
 import com.liferay.portal.security.exportimport.UserOperation;
-import com.liferay.portal.security.ldap.GroupConverterKeys;
 import com.liferay.portal.security.ldap.LDAPSettingsUtil;
 import com.liferay.portal.security.ldap.Modifications;
 import com.liferay.portal.security.ldap.PortalLDAPUtil;
